refactor(cli): split html_report.rs into sections/styling modules
html_report.rs measured 1418 production SLOC (gate, tests excluded),
over the 800-line self-scan limit. Extract its two cohesive halves
into an html_report/ subdirectory, leaving html_report.rs holding the
shared table primitives and the top-level report generators:

  - html_report/styling.rs  escaping, slugify, the language palette,
                            the inline CSS/JS chrome, <head>/<tail>
  - html_report/sections.rs the per-section renderers (global summary,
                            overview, per-language, hotspot, legend,
                            actionable-summary) and their row/total types

Pure code movement: bodies verbatim, no logic change. The foo.rs +
foo/ idiom keeps html_report.rs in place so its two `#[path]` test
includes (validators + the inline test module) stay valid; both are
unchanged. write_table_core and the write_table* family stay in
html_report.rs (shared by the section renderers and vcs_report).
Each submodule uses `use super::*` and html_report.rs re-exports both
via `pub(crate) use {sections,styling}::*`, so every `crate::html_report::X`
call site (commands, vcs_report, provenance, markdown_report) resolves
unchanged. Visibility was widened to pub(crate) only where an item is
reached across the new boundary; purely file-local helpers stay private.

Baseline refreshed: html_report.rs's stale entries re-anchored; the
two write_table* / write_language_section entries move to their new
files.

html_report.rs 1418 -> 334 production SLOC; sections.rs 729,
styling.rs 375 — all under 800.
