+++ big-code-analysis-cli/src/html_report.rs
@@ -42,441 +42,11 @@ use crate::markdown_report::{
     sloc_weighted_avg_mi, thousands,
 };
 
-/// HTML-escape a string for safe interpolation into element text or
-/// double-quoted attribute values. Returns a borrowed `Cow` when the
-/// input is already safe so the common case (most metric column names,
-/// well-formed paths) allocates nothing.
-pub(crate) fn escape_html(s: &str) -> Cow<'_, str> {
-    let needs_escape = s
-        .bytes()
-        .any(|b| matches!(b, b'&' | b'<' | b'>' | b'"' | b'\''));
-    if !needs_escape {
-        return Cow::Borrowed(s);
-    }
-    let mut out = String::with_capacity(s.len() + 8);
-    for ch in s.chars() {
-        match ch {
-            '&' => out.push_str("&amp;"),
-            '<' => out.push_str("&lt;"),
-            '>' => out.push_str("&gt;"),
-            '"' => out.push_str("&quot;"),
-            '\'' => out.push_str("&#39;"),
-            _ => out.push(ch),
-        }
-    }
-    Cow::Owned(out)
-}
+mod sections;
+mod styling;
 
-/// Slugify a heading basis into an HTML `id`/fragment-safe token:
-/// lowercase ASCII, every run of non-`[a-z0-9]` characters collapsed to a
-/// single `-`, with leading/trailing hyphens trimmed. Built from the raw
-/// language *slug* (`"cpp"`, `"csharp"`) or a section title, never from a
-/// display name carrying HTML-special punctuation — so `C++` deep-links to
-/// `#cpp`, not a broken `#c++` fragment (issue #622).
-///
-/// An empty or all-separator basis yields `"section"` so the id is never
-/// the empty string (which is not a valid fragment target).
-fn slugify(basis: &str) -> String {
-    let mut slug = String::with_capacity(basis.len());
-    let mut prev_dash = false;
-    for ch in basis.chars() {
-        if ch.is_ascii_alphanumeric() {
-            slug.push(ch.to_ascii_lowercase());
-            prev_dash = false;
-        } else if !prev_dash {
-            slug.push('-');
-            prev_dash = true;
-        }
-    }
-    let trimmed = slug.trim_matches('-');
-    if trimmed.is_empty() {
-        "section".to_owned()
-    } else {
-        trimmed.to_owned()
-    }
-}
-
-/// One `h2` section in the table of contents plus the `h3` subsections
-/// nested under it (in document order). `(display_text, id)` for the
-/// section, then a child list of the same for each `h3` that followed it
-/// before the next `h2`.
-#[derive(Default)]
-struct TocSection {
-    text: String,
-    id: String,
-    children: Vec<(String, String)>,
-}
-
-/// Collects heading ids during body rendering: assigns a unique, slug-based
-/// `id` to every `h2`/`h3` (deduplicating collisions with a `-2`, `-3`, …
-/// suffix so two languages' "Summary" headings never share a fragment), and
-/// records the `h2` sections plus their nested `h3` subsections for the
-/// table-of-contents `<nav>`.
-///
-/// The TOC nests each language's `h3` hotspot subsections under its `h2`
-/// entry as a collapsible list (issue #685), reusing the per-language-unique
-/// ids `unique_id` mints so a reader can jump straight to one hotspot table.
-/// (The original `h2`-only "compact nav" — issue #622 — abandoned a reader to
-/// scrolling once per-language sections grew this many subsections.)
-#[derive(Default)]
-pub(crate) struct Headings {
-    /// Slug -> number of times already emitted, for `-N` de-duplication.
-    seen: HashMap<String, usize>,
-    /// One entry per `h2`, in document order, each carrying its `h3` children.
-    toc: Vec<TocSection>,
-}
-
-impl Headings {
-    /// Reserve a unique id for `basis`, recording the collision count so a
-    /// repeated basis (e.g. each language's "Summary") gets `-2`, `-3`, ….
-    fn unique_id(&mut self, basis: &str) -> String {
-        let slug = slugify(basis);
-        let count = self.seen.entry(slug.clone()).or_insert(0);
-        *count += 1;
-        if *count == 1 {
-            slug
-        } else {
-            format!("{slug}-{count}")
-        }
-    }
-
-    /// Record a TOC `h3` child under the most recent `h2`. A stray `h3` with
-    /// no preceding `h2` is dropped from the TOC (it cannot nest under
-    /// anything) — this never happens in practice, since every `h3` the
-    /// renderers emit sits inside a language `<section>` whose `h2` came
-    /// first.
-    fn push_child(&mut self, text: String, id: String) {
-        if let Some(section) = self.toc.last_mut() {
-            section.children.push((text, id));
-        }
-    }
-
-    /// Emit `<h2 id="…">text</h2>` and open a new TOC section. `id_basis` is
-    /// the slug source (the raw language slug or a section title);
-    /// `display_text` is the already-`escape_html`-ed visible text.
-    pub(crate) fn emit_h2(&mut self, out: &mut String, id_basis: &str, display_text: &str) {
-        let id = self.unique_id(id_basis);
-        let _ = write!(out, "<h2 id=\"{id}\">{display_text}</h2>");
-        self.toc.push(TocSection {
-            text: display_text.to_owned(),
-            id,
-            children: Vec::new(),
-        });
-    }
-
-    /// Emit `<h3 id="…">text</h3>` and nest it under the current `h2` in the
-    /// TOC. `id_basis` is the slug source; `display_text` is already
-    /// `escape_html`-ed.
-    pub(crate) fn emit_h3(&mut self, out: &mut String, id_basis: &str, display_text: &str) {
-        let id = self.unique_id(id_basis);
-        let _ = writeln!(out, "<h3 id=\"{id}\">{display_text}</h3>");
-        self.push_child(display_text.to_owned(), id);
-    }
-
-    /// Emit `<hN id="…">text</hN>` at a caller-chosen `level` (used by the
-    /// VCS report, whose bus-factor subsection sits at `h2` on the
-    /// standalone page and `h3` when embedded). A level-2 heading opens a new
-    /// TOC section; a level-3 heading nests under the current one, matching
-    /// `emit_h2`/`emit_h3`. `display_text` is already `escape_html`-ed.
-    pub(crate) fn emit_heading(
-        &mut self,
-        out: &mut String,
-        level: usize,
-        id_basis: &str,
-        display_text: &str,
-    ) {
-        let id = self.unique_id(id_basis);
-        let _ = writeln!(out, "<h{level} id=\"{id}\">{display_text}</h{level}>");
-        if level == 2 {
-            self.toc.push(TocSection {
-                text: display_text.to_owned(),
-                id,
-                children: Vec::new(),
-            });
-        } else if level == 3 {
-            self.push_child(display_text.to_owned(), id);
-        }
-    }
-}
-
-/// Emit the table-of-contents `<nav>` linking each collected `h2` section,
-/// with its `h3` subsections nested in a collapsible `<details>` so a reader
-/// can jump straight to one hotspot table (issue #685). Renders nothing when
-/// there are no sections (the empty-walk report). All `text` is already
-/// `escape_html`-ed and every `id` is slug-safe ASCII, so each entry resolves
-/// to a real anchor on the page.
-fn write_toc(out: &mut String, toc: &[TocSection]) {
-    if toc.is_empty() {
-        return;
-    }
-    let _ = out.write_str("<nav class=\"toc\" aria-label=\"Sections\">\n<ul>\n");
-    for section in toc {
-        if section.children.is_empty() {
-            let _ = writeln!(
-                out,
-                "<li><a href=\"#{}\">{}</a></li>",
-                section.id, section.text
-            );
-        } else {
-            // A `<details>` keeps the nested list collapsed by default so the
-            // nav stays compact, yet expandable — robust across print, mobile,
-            // and narrow viewports (issue #685).
-            let _ = out.write_str("<li><details>\n<summary>");
-            let _ = write!(out, "<a href=\"#{}\">{}</a>", section.id, section.text);
-            let _ = out.write_str("</summary>\n<ul>\n");
-            for (text, id) in &section.children {
-                let _ = writeln!(out, "<li><a href=\"#{id}\">{text}</a></li>");
-            }
-            let _ = out.write_str("</ul>\n</details></li>\n");
-        }
-    }
-    let _ = out.write_str("</ul>\n</nav>\n");
-}
-
-const INLINE_CSS: &str = "\
-body{font-family:-apple-system,BlinkMacSystemFont,Segoe UI,Roboto,sans-serif;\
-margin:1.5rem;color:#222;background:#fafafa}\
-h1{font-size:1.4rem;margin:0 0 0.5rem}\
-h2{font-size:1.15rem;margin:1.5rem 0 0.5rem;\
-border-bottom:1px solid #ccc;padding-bottom:0.25rem}\
-h3{font-size:1rem;margin:1rem 0 0.4rem;color:#444}\
-section{margin-top:2rem}\
-section.lang-section{padding:0.5rem 1rem;border-radius:4px;\
-border-left:3px solid rgba(127,127,127,0.35)}\
-section.lang-section>h2{margin-top:0.25rem}\
-section.lang-rust{background:rgba(222,128,82,0.08);border-left-color:rgba(222,128,82,0.55)}\
-section.lang-python{background:rgba(58,118,196,0.08);border-left-color:rgba(58,118,196,0.55)}\
-section.lang-javascript{background:rgba(229,202,71,0.10);border-left-color:rgba(229,202,71,0.65)}\
-section.lang-typescript{background:rgba(46,116,194,0.08);border-left-color:rgba(46,116,194,0.55)}\
-section.lang-java{background:rgba(196,69,60,0.08);border-left-color:rgba(196,69,60,0.55)}\
-section.lang-kotlin{background:rgba(193,71,167,0.08);border-left-color:rgba(193,71,167,0.55)}\
-section.lang-go{background:rgba(0,173,181,0.08);border-left-color:rgba(0,173,181,0.55)}\
-section.lang-cpp{background:rgba(120,80,180,0.08);border-left-color:rgba(120,80,180,0.55)}\
-section.lang-csharp{background:rgba(83,150,80,0.08);border-left-color:rgba(83,150,80,0.55)}\
-section.lang-php{background:rgba(98,113,178,0.08);border-left-color:rgba(98,113,178,0.55)}\
-section.lang-bash{background:rgba(96,128,96,0.08);border-left-color:rgba(96,128,96,0.55)}\
-section.lang-perl{background:rgba(180,120,60,0.08);border-left-color:rgba(180,120,60,0.55)}\
-section.lang-lua{background:rgba(0,86,180,0.08);border-left-color:rgba(0,86,180,0.55)}\
-section.lang-tcl{background:rgba(160,90,140,0.08);border-left-color:rgba(160,90,140,0.55)}\
-section.lang-irules{background:rgba(200,70,120,0.08);border-left-color:rgba(200,70,120,0.55)}\
-section.lang-ruby{background:rgba(204,52,45,0.08);border-left-color:rgba(204,52,45,0.55)}\
-section.lang-elixir{background:rgba(110,73,153,0.08);border-left-color:rgba(110,73,153,0.55)}\
-section.lang-other{background:rgba(127,127,127,0.06);border-left-color:rgba(127,127,127,0.45)}\
-@media (prefers-color-scheme:dark){\
-section.lang-rust{background:rgba(222,128,82,0.16)}\
-section.lang-python{background:rgba(58,118,196,0.18)}\
-section.lang-javascript{background:rgba(229,202,71,0.16)}\
-section.lang-typescript{background:rgba(46,116,194,0.18)}\
-section.lang-java{background:rgba(196,69,60,0.18)}\
-section.lang-kotlin{background:rgba(193,71,167,0.18)}\
-section.lang-go{background:rgba(0,173,181,0.18)}\
-section.lang-cpp{background:rgba(120,80,180,0.20)}\
-section.lang-csharp{background:rgba(83,150,80,0.18)}\
-section.lang-php{background:rgba(98,113,178,0.20)}\
-section.lang-bash{background:rgba(96,128,96,0.18)}\
-section.lang-perl{background:rgba(180,120,60,0.18)}\
-section.lang-lua{background:rgba(0,86,180,0.20)}\
-section.lang-tcl{background:rgba(160,90,140,0.20)}\
-section.lang-irules{background:rgba(200,70,120,0.20)}\
-section.lang-ruby{background:rgba(204,52,45,0.18)}\
-section.lang-elixir{background:rgba(110,73,153,0.20)}\
-section.lang-other{background:rgba(200,200,200,0.10)}\
-}\
-.summary{font-size:0.9rem;color:#444;margin-bottom:0.5rem}\
-.summary strong{color:#222}\
-.summary p{margin:0.2rem 0}\
-.note{font-size:0.85rem;color:#555;margin:0.4rem 0}\
-nav.toc{font-size:0.9rem;margin:0.8rem 0;padding:0.5rem 0.8rem;\
-background:#fff;border:1px solid #e5e5e5;border-radius:4px;\
-box-shadow:0 1px 2px rgba(0,0,0,0.06)}\
-nav.toc ul{margin:0.2rem 0 0.2rem 1.2rem}\
-nav.toc li{font-size:0.9rem}\
-nav.toc details>summary{cursor:pointer;list-style:revert}\
-nav.toc details ul{margin-left:1rem}\
-div.global-actionable{margin:0.8rem 0}\
-div.global-actionable ul{margin:0.2rem 0 0.2rem 1.2rem}\
-footer.provenance{margin-top:2rem;padding-top:0.5rem;border-top:1px solid #ccc;\
-font-size:0.8rem;color:#666}\
-p.sort-hint{font-size:0.85rem;color:#555;margin:0.4rem 0;font-style:italic}\
-@media (prefers-color-scheme:dark){nav.toc{background:#1e1e1e;border-color:#333}}\
-details.legend{font-size:0.85rem;color:#444;margin:0.8rem 0}\
-details.legend summary{cursor:pointer;font-weight:600;color:#222}\
-details.legend dl{margin:0.4rem 0 0 0}\
-details.legend dt{font-weight:600;margin-top:0.3rem}\
-details.legend dd{margin:0 0 0 1rem;color:#555}\
-ul{margin:0.4rem 0 0.4rem 1.2rem;padding:0}\
-li{margin:0.15rem 0;font-size:0.9rem}\
-div.table-wrap{overflow-x:auto;margin-bottom:0.5rem}\
-table.hotspot{border-collapse:collapse;width:100%;font-size:0.85rem;\
-background:#fff;box-shadow:0 1px 2px rgba(0,0,0,0.06);margin-bottom:0.5rem}\
-table.hotspot th,table.hotspot td{padding:0.4rem 0.6rem;\
-border-bottom:1px solid #e5e5e5;text-align:left;white-space:nowrap}\
-table.hotspot th{background:#f0f0f0;cursor:pointer;user-select:none;\
-font-weight:600}\
-table.hotspot th:hover{background:#e5e5e5}\
-table.hotspot th[aria-sort=ascending]::after{content:\" \\2191\"}\
-table.hotspot th[aria-sort=descending]::after{content:\" \\2193\"}\
-table.hotspot tr:nth-child(even) td{background:#fafafa}\
-table.hotspot td.numeric{text-align:right;font-variant-numeric:tabular-nums}\
-table.hotspot td.risk-heat-0,table.hotspot td.risk-heat-1,\
-table.hotspot td.risk-heat-2,table.hotspot td.risk-heat-3,\
-table.hotspot td.risk-heat-4{font-weight:600}\
-table.hotspot tr:nth-child(even) td.risk-heat-0,table.hotspot td.risk-heat-0\
-{background:#f4b8b0;color:#222}\
-table.hotspot tr:nth-child(even) td.risk-heat-1,table.hotspot td.risk-heat-1\
-{background:#f6cba3;color:#222}\
-table.hotspot tr:nth-child(even) td.risk-heat-2,table.hotspot td.risk-heat-2\
-{background:#f1e3a3;color:#222}\
-table.hotspot tr:nth-child(even) td.risk-heat-3,table.hotspot td.risk-heat-3\
-{background:#cfe6b0;color:#222}\
-table.hotspot tr:nth-child(even) td.risk-heat-4,table.hotspot td.risk-heat-4\
-{background:#bfe3c0;color:#222}\
-@media (prefers-color-scheme:dark){\
-body{color:#e0e0e0;background:#121212}\
-a{color:#6aa3e0}\
-h2{border-bottom-color:#333}\
-h3{color:#bbb}\
-.summary{color:#bbb}\
-.summary strong{color:#e0e0e0}\
-.note{color:#999}\
-footer.provenance{color:#888;border-top-color:#333}\
-p.sort-hint{color:#999}\
-details.legend{color:#bbb}\
-details.legend summary{color:#e0e0e0}\
-details.legend dd{color:#999}\
-table.hotspot{background:#1e1e1e}\
-table.hotspot th,table.hotspot td{border-bottom-color:#333}\
-table.hotspot th{background:#2a2a2a}\
-table.hotspot th:hover{background:#383838}\
-table.hotspot tr:nth-child(even) td{background:#232323}\
-table.hotspot tr:nth-child(even) td.risk-heat-0,table.hotspot td.risk-heat-0\
-{background:#7a1f17;color:#f0f0f0}\
-table.hotspot tr:nth-child(even) td.risk-heat-1,table.hotspot td.risk-heat-1\
-{background:#7a4a12;color:#f0f0f0}\
-table.hotspot tr:nth-child(even) td.risk-heat-2,table.hotspot td.risk-heat-2\
-{background:#6b5e12;color:#f0f0f0}\
-table.hotspot tr:nth-child(even) td.risk-heat-3,table.hotspot td.risk-heat-3\
-{background:#33591f;color:#f0f0f0}\
-table.hotspot tr:nth-child(even) td.risk-heat-4,table.hotspot td.risk-heat-4\
-{background:#1f5a2e;color:#f0f0f0}\
-}\
-";
-
-/// `LANG::name()` -> CSS class suffix table. The renderer uses
-/// every entry here; `language_palette_classes_have_css` walks
-/// [`INLINE_CSS`] to confirm both the light and dark rules exist for
-/// each suffix, so adding a row without the matching CSS fails the
-/// suite. `"other"` is the neutral fallback for any name not listed.
-///
-/// Names match production output of [`big_code_analysis::LANG::name`]
-/// (see `src/langs.rs`), which since #540 is the canonical lowercase
-/// slug for every variant (`"cpp"`, `"csharp"`, `"tsx"`). `LANG::Tsx`
-/// (`"tsx"`) reuses the `"typescript"` tint (it is TypeScript + JSX),
-/// and the Mozilla-fork `"mozjs"` reuses the `"javascript"` tint (it
-/// is JavaScript, just a different grammar), and likewise `"mozcpp"`
-/// reuses the `"cpp"` tint — none needs its own CSS rule.
-const LANGUAGE_PALETTE: &[(&str, &str)] = &[
-    ("rust", "rust"),
-    ("python", "python"),
-    ("javascript", "javascript"),
-    ("mozjs", "javascript"),
-    ("typescript", "typescript"),
-    ("tsx", "typescript"),
-    ("java", "java"),
-    ("kotlin", "kotlin"),
-    ("go", "go"),
-    ("cpp", "cpp"),
-    ("c", "cpp"),
-    ("mozcpp", "cpp"),
-    ("csharp", "csharp"),
-    ("php", "php"),
-    ("bash", "bash"),
-    ("perl", "perl"),
-    ("lua", "lua"),
-    ("tcl", "tcl"),
-    ("irules", "irules"),
-    ("ruby", "ruby"),
-    ("elixir", "elixir"),
-];
-
-fn language_palette_slug(lang_name: &str) -> &'static str {
-    LANGUAGE_PALETTE
-        .iter()
-        .find_map(|&(name, slug)| (name == lang_name).then_some(slug))
-        .unwrap_or("other")
-}
-
-const INLINE_JS: &str = "\
-(function(){\
-function num(s){return s===''?Number.POSITIVE_INFINITY:parseFloat(s.replace(/,/g,''));}\
-document.querySelectorAll('table.hotspot').forEach(function(table){\
-var headers=table.querySelectorAll('thead th');\
-headers.forEach(function(th,idx){\
-th.addEventListener('click',function(){sort(table,idx,th);});\
-});\
-});\
-function sort(tbl,idx,th){\
-var tbody=tbl.tBodies[0];\
-if(!tbody)return;\
-var rows=Array.prototype.slice.call(tbody.rows);\
-var numeric=th.dataset.numeric==='1';\
-var dir=th.getAttribute('aria-sort')==='ascending'?'descending':'ascending';\
-tbl.querySelectorAll('thead th').forEach(function(h){h.removeAttribute('aria-sort');});\
-th.setAttribute('aria-sort',dir);\
-var sign=dir==='ascending'?1:-1;\
-rows.sort(function(a,b){\
-var av=a.cells[idx].textContent;\
-var bv=b.cells[idx].textContent;\
-if(numeric){\
-var an=num(av);\
-var bn=num(bv);\
-if(an<bn)return -1*sign;\
-if(an>bn)return 1*sign;\
-return 0;\
-}\
-return av.localeCompare(bv)*sign;\
-});\
-rows.forEach(function(r){tbody.appendChild(r);});\
-}\
-})();\
-";
-
-/// Tooltips for the Per-language overview table's columns, which are NOT
-/// part of the hotspot `SPECS` (they are per-language aggregates, not
-/// per-function rows). The averaged CC / Cognitive columns reuse the
-/// shared metric definitions from [`crate::markdown_report::hotspot`] so
-/// "Avg CC" and the hotspot "CC" column can never describe the same metric
-/// differently. "Avg MI" is the exception: it is SLOC-weighted and derived
-/// from the *unclamped* MI, so it uses [`hotspot::AVG_MI_TOOLTIP`] rather
-/// than the per-file `MI_TOOLTIP` (issue #725).
-/// "Files" here means *source files analysed*; the bus-factor table's
-/// like-named column means files-per-directory and supplies its own
-/// tooltip (issue #610), so this entry never leaks onto it.
-const AST_OVERVIEW_TOOLTIPS: &[(&str, &str)] = &[
-    ("SLOC", hotspot::SLOC_TOOLTIP),
-    ("Avg MI", hotspot::AVG_MI_TOOLTIP),
-    ("Avg CC", hotspot::CC_TOOLTIP),
-    ("Avg Cognitive", hotspot::COGNITIVE_TOOLTIP),
-    ("Functions", "Number of functions and methods analysed."),
-    ("Files", "Number of source files analysed."),
-];
-
-/// Plain-English tooltip for a metric column header, or `None` when the
-/// header names a non-metric dimension (file, function, class, line,
-/// language) or a column whose table supplies its own tooltips
-/// out-of-band (the hotspot and change-history tables, which pass their
-/// spec tooltips through [`write_table_with_tooltips`]).
-///
-/// This lookup serves only the tables that call the plain [`write_table`]
-/// (the Per-language overview), so it consults [`AST_OVERVIEW_TOOLTIPS`].
-/// The hotspot and VCS columns own their definitions on
-/// [`crate::markdown_report::hotspot::Column`] /
-/// [`crate::vcs_report::VcsColumn`] and never route through here.
-fn header_tooltip(header: &str) -> Option<&'static str> {
-    AST_OVERVIEW_TOOLTIPS
-        .iter()
-        .find_map(|&(name, tip)| (name == header).then_some(tip))
-}
+pub(crate) use sections::*;
+pub(crate) use styling::*;
 
 /// The column a table is pre-ranked by, and its sort direction's
 /// `aria-sort` value (`"ascending"` / `"descending"`). The HTML renderer
@@ -655,258 +225,6 @@ fn write_table_core<'t>(
     let _ = out.write_str("</tbody>\n</table>\n</div>\n");
 }
 
-/// Render a hotspot table from the shared [`Column`] descriptors. Builds the
-/// parallel `headers`/`aligns`/`rows` arrays and delegates to [`write_table`],
-/// the single source of truth for table bytes and escaping. HTML escapes
-/// every cell uniformly, so the raw [`Cell`] payload (regardless of kind) is
-/// handed to `write_table`, whose `escape_html` runs exactly once per cell.
-fn write_hotspot_table(out: &mut String, spec: &HotspotSpec, entries: &[&FunctionSummary]) {
-    let columns = spec.columns;
-    let mut headers = Vec::with_capacity(columns.len());
-    let mut aligns = Vec::with_capacity(columns.len());
-    let mut tooltips = Vec::with_capacity(columns.len());
-    for col in columns {
-        headers.push(col.header);
-        aligns.push(col.align);
-        tooltips.push(col.tooltip);
-    }
-    let mut rows: Vec<Vec<String>> = Vec::with_capacity(entries.len());
-    for s in entries {
-        let mut row: Vec<String> = Vec::with_capacity(columns.len());
-        for col in columns {
-            let (Cell::Name(text) | Cell::Path(text) | Cell::Num(text)) = (col.cell)(s);
-            row.push(text);
-        }
-        rows.push(row);
-    }
-    // The column spec carries the authoritative tooltip; pass it
-    // positionally so each header's `title=` matches the legend exactly.
-    // The spec also names the pre-ranked column and its direction, so the
-    // table announces its initial sort order without renderer guesswork.
-    let ranked = RankedColumn {
-        index: spec.rank_col,
-        dir: spec.dir.aria_sort(),
-    };
-    write_table_with_tooltips(out, &headers, &aligns, &tooltips, Some(ranked), &rows);
-}
-
-/// Per-language grouping of summaries, keyed by `LANG::name()` and
-/// ordered alphabetically (so the report sections are deterministic).
-type LangGroups<'a> = BTreeMap<&'a str, Vec<&'a FunctionSummary>>;
-
-/// Group summaries by language name. The `BTreeMap` ordering drives the
-/// alphabetical section order asserted by
-/// `two_language_well_formed_and_alphabetical`.
-fn group_by_language(summaries: &[FunctionSummary]) -> LangGroups<'_> {
-    let mut map = LangGroups::new();
-    for s in summaries {
-        map.entry(s.language.name()).or_default().push(s);
-    }
-    map
-}
-
-/// Comment lines as a percentage of source lines, guarding the
-/// zero-SLOC case. Shared by the global and per-language roll-ups so the
-/// formula lives in one place.
-fn comment_ratio_percent(sloc: usize, cloc: usize) -> f64 {
-    if sloc > 0 {
-        (cloc as f64 / sloc as f64) * 100.0
-    } else {
-        0.0
-    }
-}
-
-/// Whole-walk roll-up shown in the global `<div class="summary">` block.
-/// Only `SpaceKind::Unit` summaries contribute file-level line counts;
-/// functions and class-likes are counted separately.
-struct GlobalTotals {
-    files: usize,
-    sloc: usize,
-    ploc: usize,
-    cloc: usize,
-    functions: usize,
-    classes: usize,
-}
-
-impl GlobalTotals {
-    fn from_summaries(summaries: &[FunctionSummary]) -> Self {
-        let mut t = Self {
-            files: 0,
-            sloc: 0,
-            ploc: 0,
-            cloc: 0,
-            functions: 0,
-            classes: 0,
-        };
-        for s in summaries {
-            match s.kind {
-                SpaceKind::Unit => {
-                    t.files += 1;
-                    t.sloc += s.sloc;
-                    t.ploc += s.ploc;
-                    t.cloc += s.cloc;
-                }
-                SpaceKind::Function => t.functions += 1,
-                _ => {}
-            }
-            if is_class_like(s.kind) {
-                t.classes += 1;
-            }
-        }
-        t
-    }
-
-    fn comment_ratio(&self) -> f64 {
-        comment_ratio_percent(self.sloc, self.cloc)
-    }
-}
-
-/// Write the shared HTML preamble (doctype, head with the inline CSS, and
-/// the opening `<h1>`). `title` fills the `<title>` element and `heading`
-/// the `<h1>`; both are escaped. Shared by the AST report and the VCS
-/// report (`crate::vcs_report`) so the two pages carry identical styling
-/// and the sortable-table CSS.
-pub(crate) fn write_html_head(out: &mut String, title: &str, heading: &str) {
-    let _ = out.write_str("<!doctype html>\n<html lang=\"en\">\n<head>\n");
-    let _ = out.write_str("<meta charset=\"utf-8\">\n");
-    // Without a viewport meta, mobile browsers render the page at desktop
-    // width zoomed out to unreadable; the report is published to GitHub Pages
-    // where mobile and split-screen reading are normal (issue #686).
-    let _ =
-        out.write_str("<meta name=\"viewport\" content=\"width=device-width, initial-scale=1\">\n");
-    let _ = writeln!(
-        out,
-        "<title>{} \u{2014} big-code-analysis</title>",
-        escape_html(title)
-    );
-    let _ = writeln!(out, "<style>{INLINE_CSS}</style>");
-    let _ = out.write_str("</head>\n<body>\n");
-    let _ = writeln!(out, "<h1>{}</h1>", escape_html(heading));
-}
-
-fn write_global_summary(out: &mut String, totals: &GlobalTotals, by_lang: &LangGroups<'_>) {
-    let languages_list: String = by_lang
-        .keys()
-        .map(|k| language_display_name(k))
-        .collect::<Vec<_>>()
-        .join(", ");
-
-    let _ = out.write_str("<div class=\"summary\">\n");
-    let _ = writeln!(
-        out,
-        "<p><strong>Files analyzed:</strong> {} <strong>Languages:</strong> {}</p>",
-        escape_html(&thousands(totals.files)),
-        escape_html(&languages_list),
-    );
-    // PLOC / Comments carry hover tooltips (the legend defines them too —
-    // issue #679); SLOC's definition already rides the hotspot SLOC column.
-    let _ = writeln!(
-        out,
-        "<p><strong>Total SLOC:</strong> {} \
-         <strong title=\"{}\">PLOC:</strong> {} \
-         <strong title=\"{}\">Comments:</strong> {}</p>",
-        escape_html(&thousands(totals.sloc)),
-        escape_html(hotspot::PLOC_TOOLTIP),
-        escape_html(&thousands(totals.ploc)),
-        escape_html(hotspot::COMMENTS_TOOLTIP),
-        escape_html(&thousands(totals.cloc)),
-    );
-    let _ = writeln!(
-        out,
-        "<p><strong>Functions/methods:</strong> {} <strong>Types:</strong> {}</p>",
-        escape_html(&thousands(totals.functions)),
-        escape_html(&thousands(totals.classes)),
-    );
-    let _ = writeln!(
-        out,
-        "<p><strong>Comment ratio:</strong> {:.1}%</p>",
-        totals.comment_ratio()
-    );
-    let _ = out.write_str("</div>\n");
-}
-
-/// Build the seven-cell per-language overview row (Files / SLOC /
-/// Functions averaged from the unit and function summaries of one
-/// language).
-fn overview_row(lang_name: &str, lang_summaries: &[&FunctionSummary]) -> Vec<String> {
-    let mut unit_count = 0usize;
-    let mut sloc = 0usize;
-    let mut mi_numerator = 0.0f64;
-    let mut func_count = 0usize;
-    let mut cc_sum = 0.0f64;
-    let mut cog_sum = 0.0f64;
-    for s in lang_summaries {
-        match s.kind {
-            SpaceKind::Unit => {
-                unit_count += 1;
-                sloc += s.sloc;
-                mi_numerator += mi_weight_numerator(s);
-            }
-            SpaceKind::Function => {
-                func_count += 1;
-                cc_sum += s.cyclomatic;
-                cog_sum += s.cognitive;
-            }
-            _ => {}
-        }
-    }
-    let avg_mi = sloc_weighted_avg_mi(mi_numerator, sloc);
-    let (avg_cc, avg_cog) = if func_count > 0 {
-        (cc_sum / func_count as f64, cog_sum / func_count as f64)
-    } else {
-        (0.0, 0.0)
-    };
-    vec![
-        language_display_name(lang_name).into_owned(),
-        thousands(unit_count),
-        thousands(sloc),
-        thousands(func_count),
-        format!("{avg_mi:.1}"),
-        format!("{avg_cc:.1}"),
-        format!("{avg_cog:.1}"),
-    ]
-}
-
-fn write_overview_table(out: &mut String, headings: &mut Headings, by_lang: &LangGroups<'_>) {
-    headings.emit_h2(out, "Per-language overview", "Per-language overview");
-    let _ = out.write_str("\n");
-    // One discoverability hint near the first table: the columns are
-    // sortable, which is otherwise invisible until a reader happens to
-    // click a header (issue #622).
-    let _ = out.write_str("<p class=\"sort-hint\">Click a column header to sort any table.</p>\n");
-    let mut rows: Vec<Vec<String>> = Vec::with_capacity(by_lang.len());
-    for (&lang_name, lang_summaries) in by_lang {
-        rows.push(overview_row(lang_name, lang_summaries));
-    }
-    write_table(
-        out,
-        &[
-            "Language",
-            "Files",
-            "SLOC",
-            "Functions",
-            "Avg MI",
-            "Avg CC",
-            "Avg Cognitive",
-        ],
-        &[
-            Align::Left,
-            Align::Right,
-            Align::Right,
-            Align::Right,
-            Align::Right,
-            Align::Right,
-            Align::Right,
-        ],
-        &rows,
-    );
-}
-
-pub(crate) fn write_html_tail(out: &mut String) {
-    let _ = writeln!(out, "<script>{INLINE_JS}</script>");
-    let _ = out.write_str("</body>\n</html>\n");
-}
-
 /// Produce a self-contained HTML quality-metrics report from the
 /// collected summaries. `top_n` controls how many entries appear in
 /// each hotspot table. Footer-free convenience wrapper used by the snapshot
@@ -1005,408 +323,6 @@ pub(crate) fn generate_html_report_with_vcs(
     out.push_str(&body);
     write_html_tail(&mut out);
     out
-}
-
-/// Render the global cross-language Actionable Summary near the TOC: the
-/// per-language counts summed over every function in the walk (issue #678).
-/// Suppression policy does not gate these counts — like the per-language
-/// Actionable Summary, this is a raw whole-codebase health signal, not a
-/// gate (issue #501). Renders the "no concerns" line when nothing clears a
-/// threshold, mirroring the per-language block.
-fn write_global_actionable_summary(
-    out: &mut String,
-    summaries: &[FunctionSummary],
-    advisory: AdvisoryThresholds,
-) {
-    let funcs: Vec<&FunctionSummary> = summaries
-        .iter()
-        .filter(|s| s.kind == SpaceKind::Function)
-        .collect();
-    let counts = advisory.count_over(&funcs);
-    let _ = out.write_str(
-        "<div class=\"summary global-actionable\">\n<p><strong>Actionable summary \
-         (all languages)</strong></p>\n",
-    );
-    // Provenance so the global roll-up's cutoffs are attributable (issue #630).
-    let _ = writeln!(
-        out,
-        "<p class=\"note\">{}</p>",
-        escape_html(advisory.provenance_line())
-    );
-    if counts.all_clear() {
-        let _ = out.write_str("<p class=\"note\">No major quality concerns detected.</p>\n");
-        let _ = out.write_str("</div>\n");
-        return;
-    }
-    write_actionable_bullets(out, counts, advisory);
-    let _ = out.write_str("</div>\n");
-}
-
-/// Emit a visible legend listing each metric column's abbreviation and its
-/// one-line definition, drawn from the same `(header, tooltip)` pairs the
-/// hover tooltips use so the two cannot drift. The `<details>` is rendered
-/// **open** — the legend's whole purpose is to survive print, mobile, and
-/// screen readers, all of which a collapsed `<details>` defeats just as the
-/// hover tooltips do (issue #679); the page bottom is cheap vertical space.
-/// Renders nothing when `entries` is empty.
-pub(crate) fn write_legend_html(out: &mut String, entries: &[(&str, &str)]) {
-    if entries.is_empty() {
-        return;
-    }
-    let _ = out.write_str("<details class=\"legend\" open>\n<summary>Legend</summary>\n<dl>\n");
-    for (header, tip) in entries {
-        // Link the term to its hosted metric chapter when one exists (issue
-        // #675); VCS / identity headers have no anchor and render bare.
-        let term = match hotspot::metric_doc_url(header) {
-            Some(url) => format!(
-                "<a href=\"{}\">{}</a>",
-                escape_html(&url),
-                escape_html(header)
-            ),
-            None => escape_html(header).into_owned(),
-        };
-        let _ = writeln!(out, "<dt>{term}</dt><dd>{}</dd>", escape_html(tip));
-    }
-    let _ = out.write_str("</dl>\n</details>\n");
-}
-
-/// Split a per-language slice into its unit (file) and function buckets
-/// in a single pass. Class-likes are intentionally dropped from both
-/// buckets — the WMC hotspot sources them straight from `entries`.
-fn partition_by_kind<'a>(
-    entries: &[&'a FunctionSummary],
-) -> (Vec<&'a FunctionSummary>, Vec<&'a FunctionSummary>) {
-    let mut units: Vec<&FunctionSummary> = Vec::with_capacity(entries.len());
-    let mut funcs: Vec<&FunctionSummary> = Vec::with_capacity(entries.len());
-    for &s in entries {
-        match s.kind {
-            SpaceKind::Unit => units.push(s),
-            SpaceKind::Function => funcs.push(s),
-            _ => {}
-        }
-    }
-    (units, funcs)
-}
-
-/// File-level roll-up backing one language's `<h3>Summary</h3>` note.
-/// Only `SpaceKind::Unit` summaries feed it.
-struct LanguageTotals {
-    files: usize,
-    sloc: usize,
-    ploc: usize,
-    cloc: usize,
-    /// SLOC-weighted numerator of the unclamped Visual Studio MI: the sum
-    /// of [`mi_weight_numerator`] over the units. Divided by `sloc` (not
-    /// `files`) to form the headline average (issue #725).
-    mi_numerator: f64,
-}
-
-impl LanguageTotals {
-    fn from_units(units: &[&FunctionSummary]) -> Self {
-        let mut t = Self {
-            files: 0,
-            sloc: 0,
-            ploc: 0,
-            cloc: 0,
-            mi_numerator: 0.0,
-        };
-        for s in units {
-            t.files += 1;
-            t.sloc += s.sloc;
-            t.ploc += s.ploc;
-            t.cloc += s.cloc;
-            t.mi_numerator += mi_weight_numerator(s);
-        }
-        t
-    }
-
-    fn comment_ratio(&self) -> f64 {
-        comment_ratio_percent(self.sloc, self.cloc)
-    }
-
-    fn avg_mi(&self) -> f64 {
-        sloc_weighted_avg_mi(self.mi_numerator, self.sloc)
-    }
-}
-
-fn write_language_header(out: &mut String, headings: &mut Headings, lang_name: &str) {
-    let display_name = language_display_name(lang_name);
-    // `slug` is sourced from `LANGUAGE_PALETTE` (or the literal "other"
-    // fallback) — always lowercase ASCII, so it is interpolated raw
-    // into the class attribute without `escape_html`.
-    let slug = language_palette_slug(lang_name);
-    let _ = write!(out, "<section class=\"lang-section lang-{slug}\">");
-    // The heading id is slug-based off the raw language name (`lang_name`,
-    // e.g. "cpp"/"csharp"), NOT the display name — so `C++` deep-links to a
-    // valid `#cpp` fragment rather than the punctuation-laden display text.
-    headings.emit_h2(out, lang_name, &escape_html(&display_name));
-    let _ = out.write_str("\n");
-}
-
-fn write_language_summary(
-    out: &mut String,
-    headings: &mut Headings,
-    id_prefix: &str,
-    units: &[&FunctionSummary],
-) {
-    let totals = LanguageTotals::from_units(units);
-    let cr = totals.comment_ratio();
-    let avg_mi = totals.avg_mi();
-    let rating = mi_rating(avg_mi);
-
-    headings.emit_h3(out, &format!("{id_prefix}-summary"), "Summary");
-    let _ = writeln!(
-        out,
-        "<p class=\"note\">Files: {} | SLOC: {} | PLOC: {} | Comment ratio: {cr:.1}%</p>",
-        escape_html(&thousands(totals.files)),
-        escape_html(&thousands(totals.sloc)),
-        escape_html(&thousands(totals.ploc)),
-    );
-    let _ = writeln!(
-        out,
-        "<p class=\"note\">{}: {avg_mi:.1} ({rating})</p>",
-        crate::markdown_report::AVG_MI_LABEL
-    );
-}
-
-/// Emit one hotspot section: an `<h3>` heading followed by the column-driven
-/// table, from a shared [`HotspotSpec`] and its already-selected `rows`. The
-/// logical title is `escape_html`-ed here defensively (a `>` in a `concept`
-/// would become `&gt;`); since no current concept carries one (see
-/// [`hotspot::HotspotTitle::render`]), it is a no-op for today's
-/// metachar-free titles.
-fn emit_html_section(
-    out: &mut String,
-    headings: &mut Headings,
-    id_prefix: &str,
-    spec: &HotspotSpec,
-    top_n: usize,
-    rows: &[&FunctionSummary],
-) {
-    let title = spec.title.render(top_n);
-    // The id basis is the language slug plus the section's *stable* basis
-    // ("<Concept> hotspots", without the `(top N by …)` clause), so the
-    // fragment (`#rust-cyclomatic-complexity-hotspots`) does not shift with
-    // `--top` (issue #677). The displayed text is the full rendered title.
-    headings.emit_h3(
-        out,
-        &format!("{id_prefix}-{}", spec.title.id_basis()),
-        &escape_html(&title),
-    );
-    write_hotspot_table(out, spec, rows);
-}
-
-/// The cyclomatic summary note under the CC hotspot table. A caption over the
-/// same suppression-filtered set the table shows (see [`hotspot::select_cc`]),
-/// matching the Markdown report; the raw, suppression-independent CC count
-/// lives in the Actionable Summary instead. When `policy` honors suppression
-/// the line is captioned `(excluding suppressed functions)` so a reader can
-/// tell the two CC figures apart (issue #616).
-fn emit_cc_note_html(
-    out: &mut String,
-    stats: &hotspot::CyclomaticStats,
-    policy: SuppressionPolicy,
-) {
-    let caption = hotspot::cc_note_caption(policy)
-        .map(|c| format!(" ({})", escape_html(c)))
-        .unwrap_or_default();
-    // Bands use the resolved advisory CC cutoff and its severe multiple
-    // (issue #630): `> 10` / `> 20` by default, shifted by a manifest
-    // `cyclomatic = N`.
-    let _ = writeln!(
-        out,
-        "<p class=\"note\">Average CC: {:.1} | Max: {:.0} | CC &gt; {:.0}: {} functions | CC &gt; {:.0}: {} functions{caption}</p>",
-        stats.avg(),
-        stats.max,
-        stats.primary_cutoff,
-        stats.over_primary,
-        stats.severe_cutoff,
-        stats.over_severe,
-    );
-}
-
-/// Emit the `<ul>` of advisory bullets shared by the per-language and global
-/// Actionable Summaries, the resolved advisory cutoff embedded in each label
-/// (issue #630). The caller has already emitted the heading / provenance /
-/// caption and confirmed `!counts.all_clear()`.
-fn write_actionable_bullets(
-    out: &mut String,
-    counts: AdvisoryCounts,
-    advisory: AdvisoryThresholds,
-) {
-    let _ = out.write_str("<ul>\n");
-    if counts.cc > 0 {
-        let _ = writeln!(
-            out,
-            "<li><strong>{}</strong> functions with CC &gt; {:.0}</li>",
-            counts.cc, advisory.cc
-        );
-    }
-    if counts.cognitive > 0 {
-        let _ = writeln!(
-            out,
-            "<li><strong>{}</strong> functions with cognitive complexity &gt; {:.0}</li>",
-            counts.cognitive, advisory.cognitive
-        );
-    }
-    if counts.sloc > 0 {
-        let _ = writeln!(
-            out,
-            "<li><strong>{}</strong> functions with SLOC &gt; {}</li>",
-            counts.sloc, advisory.sloc
-        );
-    }
-    if counts.nargs > 0 {
-        let _ = writeln!(
-            out,
-            "<li><strong>{}</strong> functions with more than {} parameters</li>",
-            counts.nargs, advisory.nargs
-        );
-    }
-    if counts.bugs > 0 {
-        let _ = writeln!(
-            out,
-            "<li><strong>{}</strong> functions with estimated Halstead bugs &gt; {:.1}</li>",
-            counts.bugs, advisory.bugs
-        );
-    }
-    let _ = out.write_str("</ul>\n");
-}
-
-fn write_actionable_summary(
-    out: &mut String,
-    headings: &mut Headings,
-    id_prefix: &str,
-    funcs: &[&FunctionSummary],
-    policy: SuppressionPolicy,
-    advisory: AdvisoryThresholds,
-) {
-    let counts = advisory.count_over(funcs);
-    headings.emit_h3(
-        out,
-        &format!("{id_prefix}-actionable-summary"),
-        "Actionable Summary",
-    );
-    // Provenance so the cutoffs are always attributable (issue #630).
-    let _ = writeln!(
-        out,
-        "<p class=\"note\">{}</p>",
-        escape_html(advisory.provenance_line())
-    );
-    let breakdown = hotspot::suppressed_metric_breakdown(funcs, policy, advisory);
-    let _ = writeln!(
-        out,
-        "<p class=\"note\">{}</p>",
-        escape_html(&hotspot::actionable_summary_caption(&breakdown))
-    );
-    if counts.all_clear() {
-        let _ = out.write_str("<p class=\"note\">No major quality concerns detected.</p>\n");
-        return;
-    }
-    write_actionable_bullets(out, counts, advisory);
-}
-
-/// Emit the section heading (`<h3>` + id) followed by the "table omitted: all
-/// N matching functions suppressed" caption, in place of a hotspot table that
-/// was rendered empty *solely because* suppression hid every matching row
-/// (`count > 0`). The heading uses the same id basis `emit_html_section` does,
-/// so deep links resolve regardless of suppression state (issue #681).
-/// Mirrors the Markdown renderer's `emit_fully_suppressed_note_md` so a
-/// summary bullet never points at a table absent from the document (issue
-/// #616). A no-op when `count == 0`.
-fn emit_fully_suppressed_note_html(
-    out: &mut String,
-    headings: &mut Headings,
-    id_prefix: &str,
-    spec: &HotspotSpec,
-    top_n: usize,
-    count: usize,
-) {
-    if count == 0 {
-        return;
-    }
-    let title = spec.title.render(top_n);
-    // Emit the section heading with the same stable id basis
-    // `emit_html_section` uses, so a fully-suppressed section keeps its place
-    // in the heading/id sequence and deep links resolve regardless of
-    // suppression state (issue #681). The omission note is the section body.
-    headings.emit_h3(
-        out,
-        &format!("{id_prefix}-{}", spec.title.id_basis()),
-        &escape_html(&title),
-    );
-    let _ = writeln!(
-        out,
-        "<p class=\"note\">{}</p>",
-        escape_html(&hotspot::fully_suppressed_caption(&title, count))
-    );
-}
-
-fn write_language_section(
-    out: &mut String,
-    headings: &mut Headings,
-    lang_name: &str,
-    entries: &[&FunctionSummary],
-    top_n: usize,
-    policy: SuppressionPolicy,
-    advisory: AdvisoryThresholds,
-) {
-    write_language_header(out, headings, lang_name);
-    // Slug-based id prefix for this language's `h3` headings, so a hotspot
-    // table deep-links to e.g. `#rust-cyclomatic-complexity-hotspots`. Built
-    // from the raw language slug, matching the `h2` section id.
-    let id_prefix = slugify(lang_name);
-    let (units, funcs) = partition_by_kind(entries);
-    write_language_summary(out, headings, &id_prefix, &units);
-
-    // The Actionable Summary leads the section (directly after Summary,
-    // before any hotspot table) so a reader sees the highest-altitude counts
-    // first (issue #678).
-    write_actionable_summary(out, headings, &id_prefix, &funcs, policy, advisory);
-
-    // Drive every hotspot section from the shared `SPECS` table so the HTML
-    // and Markdown reports cannot diverge in membership/order/suppression.
-    // WMC draws from the full slice, MI from units, the rest from functions.
-    for spec in SPECS {
-        let base: &[&FunctionSummary] = match spec.source {
-            Source::Units => &units,
-            Source::Funcs => &funcs,
-            Source::All => entries,
-        };
-        if spec.cc_note {
-            let (rows, stats) = hotspot::select_cc(spec, base, top_n, policy, advisory);
-            if rows.is_empty() {
-                // Mirror the non-CC branch: a CC table emptied purely by
-                // suppression earns the same "table omitted" caption, or the
-                // Actionable Summary's raw CC bullets would dangle (#616).
-                let suppressed = hotspot::fully_suppressed_count(spec, base, policy, advisory);
-                emit_fully_suppressed_note_html(out, headings, &id_prefix, spec, top_n, suppressed);
-            } else {
-                emit_html_section(out, headings, &id_prefix, spec, top_n, &rows);
-                emit_cc_note_html(out, &stats, policy);
-            }
-        } else {
-            let rows = hotspot::select(spec, base, top_n, policy, advisory);
-            if rows.is_empty() {
-                // An empty table here is either a genuinely-absent metric or a
-                // table whose every matching row was suppressed; only the
-                // latter earns a caption so a summary bullet never dangles.
-                let suppressed = hotspot::fully_suppressed_count(spec, base, policy, advisory);
-                emit_fully_suppressed_note_html(out, headings, &id_prefix, spec, top_n, suppressed);
-            } else {
-                emit_html_section(out, headings, &id_prefix, spec, top_n, &rows);
-                if spec.mi_note {
-                    let _ = writeln!(
-                        out,
-                        "<p class=\"note\">{}</p>",
-                        escape_html(hotspot::MI_NOTE)
-                    );
-                }
-            }
-        }
-    }
-
-    let _ = out.write_str("</section>\n");
 }
 
 // Pull in the same `quick-xml`-driven well-formedness walker the
